@@ -11,25 +11,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.2.5",
-    "@angular/common": "~8.2.5",
-    "@angular/compiler": "~8.2.5",
-    "@angular/core": "~8.2.5",
-    "@angular/forms": "~8.2.5",
-    "@angular/platform-browser": "~8.2.5",
-    "@angular/platform-browser-dynamic": "~8.2.5",
-    "@angular/router": "~8.2.5",
+    "@angular/animations": "~9.0.0-next.5",
+    "@angular/common": "~9.0.0-next.5",
+    "@angular/compiler": "~9.0.0-next.5",
+    "@angular/core": "~9.0.0-next.5",
+    "@angular/forms": "~9.0.0-next.5",
+    "@angular/platform-browser": "~9.0.0-next.5",
+    "@angular/platform-browser-dynamic": "~9.0.0-next.5",
+    "@angular/router": "~9.0.0-next.5",
     "@ng-bootstrap/ng-bootstrap": "^5.1.1",
     "bootstrap": "^4.3.1",
     "rxjs": "~6.5.3",
     "tslib": "^1.10.0",
-    "zone.js": "~0.9.1"
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.802.2",
-    "@angular/cli": "~8.2.2",
-    "@angular/compiler-cli": "~8.2.5",
-    "@angular/language-service": "~8.2.5",
+    "@angular-devkit/build-angular": "~0.803.3",
+    "@angular/cli": "~8.3.3",
+    "@angular/compiler-cli": "~9.0.0-next.5",
+    "@angular/language-service": "~9.0.0-next.5",
     "@types/node": "~8.9.4",
     "@types/jasmine": "~3.3.8",
     "@types/jasminewd2": "~2.0.3",
